// src/types/converter.ts
export type ConverterTool =
  | 'converterList'
  | 'emiCalculatorList'
  | 'financialPlannerList'
  | 'investmentList'
  | 'otherCalculatorList'
  | 'currencyConverter'
  | 'cryptoConverter'
  | 'customRate'
  | 'currencyList'
  | 'chart'
  | 'settings'
  | 'loanComparison'
  | 'loanAnalysis'
  | 'homeAffordability'
  | 'savingsGoal'
  | 'fixedDeposit'
  | 'recurringDeposit'
  | 'sipCalculator'
  | 'returnOnInvestment'
  | 'creditCardPayoff'
  | 'creditCardMinPayment'
  | 'breakEvenSellPrice'
  | 'compoundInterest';